+++ requests.go
@@ -1,12 +1,12 @@
 package MateBotSDKGo
 
 type simpleID struct {
-	ID uint `json:"id"`
+	ID uint64 `json:"id"`
 }
 
 type newCallback struct {
 	Url           string `json:"url"`
-	ApplicationID uint   `json:"application_id"`
+	ApplicationID uint64 `json:"application_id"`
 	SharedSecret  string `json:"shared_secret"`
 }
 
@@ -15,8 +15,8 @@ type newUser struct {
 }
 
 type newAlias struct {
-	UserID        uint   `json:"user_id"`
-	ApplicationID uint   `json:"application_id"`
+	UserID        uint64 `json:"user_id"`
+	ApplicationID uint64 `json:"application_id"`
 	Username      string `json:"username"`
 	Confirmed     bool   `json:"confirmed"`
 }
@@ -24,18 +24,18 @@ type newAlias struct {
 type newTransaction struct {
 	Sender   any    `json:"sender"`
 	Receiver any    `json:"receiver"`
-	Amount   uint   `json:"amount"`
+	Amount   uint64 `json:"amount"`
 	Reason   string `json:"reason"`
 }
 
 type newConsumption struct {
 	User       any    `json:"user"`
-	Amount     uint   `json:"amount"`
+	Amount     uint64 `json:"amount"`
 	Consumable string `json:"consumable"`
 }
 
 type newCommunism struct {
-	Amount      uint   `json:"amount"`
+	Amount      uint64 `json:"amount"`
 	Description string `json:"description"`
 	Creator     any    `json:"creator"`
 }
@@ -47,20 +47,20 @@ type newPoll struct {
 }
 
 type newRefund struct {
-	Amount      uint   `json:"amount"`
+	Amount      uint64 `json:"amount"`
 	Description string `json:"description"`
 	Creator     any    `json:"creator"`
 }
 
 type newVote struct {
-	User     any  `json:"user"`
-	BallotID uint `json:"ballot_id"`
-	Vote     bool `json:"vote"`
+	User     any    `json:"user"`
+	BallotID uint64 `json:"ballot_id"`
+	Vote     bool   `json:"vote"`
 }
 
 type issuerID struct {
-	ID     uint `json:"id"`
-	Issuer any  `json:"issuer"`
+	ID     uint64 `json:"id"`
+	Issuer any    `json:"issuer"`
 }
 
 type userPrivilegeDrop struct {
@@ -80,6 +80,6 @@ type voucherUpdate struct {
 }
 
 type communismParticipationUpdate struct {
-	ID   uint `json:"id"`
-	User any  `json:"user"`
+	ID   uint64 `json:"id"`
+	User any    `json:"user"`
 }
